feat: Includes first working version of MPC
This commit includes the first working version of the MPC algorithm.

It uses a conservative slow speed to sucessfully complete a lap around
the test track.

Visualization tools are also included in this commit and the algo will
display the line that the MPC should follow (in yellow) and the next N
points as calculated by the MPC (in green).

Performance is satisfactory to pass the rubric points, but far from
ideal for a more realistic application.

Next steps:
- Fine tune hyperparameters to reduce jerkyness and increase speed